--- .pypric
@@ -1,3 +1,0 @@
-[pypi]
-  username = __token__
-  password = pypi-AgEIcHlwaS5vcmcCJDYzM2YxOWQwLTZmNGQtNDAwOS1hNzQwLWFjMzUzNzRjMTZjYQACKlszLCI5ZDBiNWRkNi1hM2Q4LTQ0OTItYWFmYy03MDRiNDE2ODNhN2QiXQAABiA9q6EcTfw_JJqG6QhnnudI9-ddpLAOXH5YjzKaGtwtew
