--- euclidean_algorithm/setup.cfg
@@ -1,1 +1,0 @@
-# Nothing
