+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "euclidean-algorithm"
-version = "0.1.1"
+version = "0.1.2"
 description = ""
 authors = ["Макс <rumyantsev1m0s1@gmail.com>"]
 license = "MIT"
@@ -9,9 +9,6 @@ readme = "README.md"
 [tool.poetry.dependencies]
 python = "^3.12"
 numba = "^0.60.0"
-setuptools = "^75.6.0"
-twine = "^6.0.1"
-travis = "^0.0.3"
 
 
 [build-system]
